Add option to launch headless
When running in a vanilla node environment on macOS, Java cannot be
loaded from the application's main thread. And I guess javascript
typically runs single threaded, so we cannot simply defer the loading.

One workaround is to use headless mode. See joeferner/node-java#21.
Another is to use Electron, where it appears the threading is different.